get options page wrapper
